v2

+++ P/Error/views.py
@@ -4,22 +4,33 @@ from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework import status
 from rest_framework_simplejwt.authentication import JWTAuthentication
-from .models import ERROR
+from .models import ERRORModel
 from .models import Comment
 from .serializers import ErrorSerializer
 from .serializers import CommentSerializer
 from rest_framework.permissions import IsAuthenticated
 
 
+'''
+user 1 ==> ADMIN
+user 2 ==> DEV
+user 3 ==> EXPERT
+
+'''
+
 
 
 @api_view(['POST'])
-@permission_classes([IsAuthenticated])
 @authentication_classes([JWTAuthentication])
+@permission_classes([IsAuthenticated])#
 def add_error(request : Request):
+    '''
+    this method will check if the user is authenticate or user has permission
+    and all users have permission to  add_errors with the solutiono and refrece of the answer.
+    '''
 
-    if not request.user.is_authenticated and request.user.has_perm('Error.add_error'):
-        return Response({"msg" : "Not Allowed"}, status=status.HTTP_401_UNAUTHORIZED) #give the message and stop.
+    if not request.user.is_authenticated or not request.user.has_perm('Error.add_errormodel'): #must be lowercase.
+        return Response({"msg" : "Not Allowed check on your authentication and Permission"}, status=status.HTTP_401_UNAUTHORIZED) #give the message and stop.
     # if not request.user.has_perm('Error.add_error'):
     #     return Response({"msg": "doesn't have permission"})
     new_error = ErrorSerializer(data=request.data)
@@ -29,6 +40,7 @@ def add_error(request : Request):
             "msg" : "Created Successfully",
             "Error" : new_error.data
         }
+
         return Response(dataResponse)
     else:
         print(new_error.errors)
@@ -39,11 +51,18 @@ def add_error(request : Request):
 
 # update error or handling something.
 @api_view(['PUT'])
-@authentication_classes([JWTAuthentication])
-@permission_classes([IsAuthenticated])
+# @authentication_classes([JWTAuthentication])
+# @permission_classes([IsAuthenticated])
   #
 def update_error(request : Request, error_id): #
-    up_error = ERROR.objects.get(id=error_id)
+
+    '''
+        this method will check if the user is authenticate
+        then so can update on the error already exsit.
+
+    '''
+
+    up_error = ERRORModel.objects.get(id=error_id)
 
     updated_error = ErrorSerializer(instance=up_error, data=request.data)
     if updated_error.is_valid():
@@ -66,11 +85,17 @@ def update_error(request : Request, error_id): #
 
 @api_view(['GET'])
 def list_error(request: Request):
+    '''
+            this method will show all the list of the errors created by users
+            and you can search by any phrase exsit in the error_description.
+
+
+        '''
     if 'search' in request.query_params:
         search_phrase = request.query_params['search']
-        errors = ERROR.objects.filter(error_desc__contains=search_phrase)
+        errors = ERRORModel.objects.filter(error_desc__contains=search_phrase)
     else:
-        errors = ERROR.objects.all()
+        errors = ERRORModel.objects.all()
 
     responseData = {
         "msg": "list of Errors",
@@ -80,37 +105,59 @@ def list_error(request: Request):
 
 
 
+@api_view(['DELETE'])
+@authentication_classes([JWTAuthentication])
+def delete_Error(request: Request, ERRORModel_id):
+    '''
+               this method will delete any error you want
+               by the error id must be authenticate.
 
-@api_view(['GET','PUT','DELETE'])
-def error_details(request,id):
 
-    try:
-       error = ERROR.objects.get(pk=id)
-    except ERROR.DoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
+    '''
 
-    if request.method == 'GET':
-       serializer = ErrorSerializer(error)
-       return Response(serializer.data)
-    elif request.method == 'PUT':
-        serializer = ErrorSerializer(error, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-    elif request.method == 'DELETE':
-            error.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
+    del_error = ERRORModel.objects.get(id=ERRORModel_id)
+    del_error.delete()
+    return Response({"msg" : "Deleted Successfully"})
+
+# @api_view(['GET','PUT','DELETE'])
+# def error_details(request,id):
+#
+#     try:
+#        error = ERRORModel.objects.get(pk=id)
+#     except ERRORModel.DoesNotExist:
+#         return Response(status=status.HTTP_404_NOT_FOUND)
+#
+#     if request.method == 'GET':
+#        serializer = ErrorSerializer(error)
+#        return Response(serializer.data)
+#     elif request.method == 'PUT':
+#         serializer = ErrorSerializer(error, data=request.data)
+#         if serializer.is_valid():
+#             serializer.save()
+#             return Response(serializer.data)
+#         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+#     elif request.method == 'DELETE':
+#             error.delete()
+#             return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 #here we start with COMMENTS
+
 @api_view(['POST'])
+@authentication_classes([JWTAuthentication])
+
+
 def addComment(request):
+    '''
+        this method will check on the user if authonticate then
+        you will be can add the comments.
+
+    '''
     if request.method == 'POST':
         serializer = CommentSerializer(data=request.data) #Validate your data with form.
         if serializer.is_valid():
             instance = serializer
-            instance.save(Commit=False) #Will save your data if valid.
+            instance.save() #Will save your data if valid.
             # data = {'success':True,'msg':"Store comment's data successfully"}
             # return Response(data)
             dataResponse = {
@@ -129,14 +176,64 @@ def addComment(request):
 
 
 @api_view(['DELETE'])
+@authentication_classes([JWTAuthentication])
+@permission_classes([IsAuthenticated])
+
 def delete_comment(request: Request, Comment_id):
+    '''
+    this function will be delete the comment
+    just Expert can delete the comment.
+    '''
+    if not request.user.is_authenticated or not request.user.has_perm('Error.delete_comment'): #must be lowercase.
+        return Response({"msg" : "Not Allowed check on your authentication and Permission"}, status=status.HTTP_401_UNAUTHORIZED)
     del_com = Comment.objects.get(id=Comment_id)
     del_com.delete()
     return Response({"msg" : "Deleted Successfully"})
 
 
+    new_error = ErrorSerializer(data=request.data)
+    if new_error.is_valid():
+        new_error.save()
+        dataResponse={
+            "msg" : "Created Successfully",
+            "Error" : new_error.data
+        }
+
+        return Response(dataResponse)
+    else:
+        print(new_error.errors)
+        dataResponse = {"msg" : "couldn't create a Error"}
+        return Response( dataResponse, status=status.HTTP_400_BAD_REQUEST)
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 @api_view(['GET'])
 def list_comment(request: Request):
+
+    '''
+    this method will show  list of the comment .
+    and search for any phrase exsit in the comment_phrase.
+
+    '''
     if 'search' in request.query_params:
         search_phrase = request.query_params['search']
         commm = Comment.objects.filter(comment__contains=search_phrase)
@@ -148,5 +245,3 @@ def list_comment(request: Request):
         "Comments": CommentSerializer(instance=commm, many=True).data
     }
     return Response(responseData)
-
-
